Only download mgba source if it's not already present

+++ mgba-test-runner/build-mgba.sh
@@ -10,7 +10,10 @@ if [[ -f libmgba-cycle.a ]]; then
 	exit 0
 fi
 
-curl -L https://github.com/mgba-emu/mgba/archive/refs/tags/${MGBA_VERSION}.tar.gz -o mgba-${MGBA_VERSION}.tar.gz
+if [[ ! -f mgba-${MGBA_VERSION}.tar.gz ]]; then
+	curl -L https://github.com/mgba-emu/mgba/archive/refs/tags/${MGBA_VERSION}.tar.gz -o mgba-${MGBA_VERSION}.tar.gz
+fi
+
 tar -xvf mgba-${MGBA_VERSION}.tar.gz
 cd mgba-${MGBA_VERSION}
 rm -rf build
